Reorganised binaryBreadthFirstSearch files
