+++ configuration.conf
@@ -30,16 +30,14 @@ vrp_basic           | N | Y | Y
 withPoints          | Y | Y | Y
 lineGraph           | Y | Y | Y
 components          | Y | Y | Y
-bellman_ford        | Y | Y | Y
+bellman_ford        | Y | Y | N
 cpp_common          | Y | N | N
 dagShortestPath     | Y | Y | Y
 chinese             | Y | Y | Y
 spanningTree        | Y | Y | Y
 mincut              | Y | Y | Y
 version             | Y | Y | Y
-breadthFirstSearch  | Y | Y | N
-binaryBreadthFirstSearch    | Y | Y | Y  
-edwardMoore         | Y | Y | N
+breadthFirstSearch  | Y | Y | Y
 #----------------------
 # SQL only directories
 #----------------------
