--- src/binaryBreadthFirstSearch/CMakeLists.txt
@@ -1,4 +1,0 @@
-ADD_LIBRARY(binaryBreadthFirstSearch OBJECT
-    binaryBreadthFirstSearch.c
-    binaryBreadthFirstSearch_driver.cpp
-    )
